move meta file handling to separate file
